# swagger.yml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Swagger Loyalty"
  contact:
    email: "apiteam@swagger.io"
host: "test.loyal-api.melifrontends.com"
basePath: "/api"
tags:
- name: "payments"
  description: "payment endpoint"
  externalDocs:
    description: "Find out more"
    url: "https://mercadolibre.com"
- name: "card_token"
  description: "token endpoint"
  externalDocs:
    description: "Find out more"
    url: "https://mercadolibre.com"
schemes:
- "http"
paths:
  /payments:
    post:
      tags:
      - "payments"
      summary: "creates payment"
      description: "Creates payment"
      operationId: "findPetsByStatus"
      produces:
      - "application/json"
      parameters:
        - in: body
          name: body
          description: Payment to create.
          schema:
            type: object
            required:
              - email
              - customer_id
              - token
              - amount
            properties:
              email:
                type: string
              customer_id:
                type: string
              token:
                type: string
              amount:
                type: number
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/PaymentResponse'
        404:
          description: "User not found"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
  /card_tokens:
    post:
      tags:
      - "card_token"
      summary: "creates card token using card_id"
      description: "creates card token for registered user"
      operationId: "createToken"
      produces:
      - "application/json"
      parameters:
        - in: body
          name: body
          description: Creates a card_token with card_id.
          schema:
            type: object
            required:
              - card_id
            properties:
              card_id:
                type: string
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/CardTokenResponse'
        404:
          description: "User not found"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
        
  /customers:
    get:
      tags:
      - "customers"
      summary: "Get mercadopago customers"
      description: "get mercadopago client customers"
      operationId: "getCustomers"
      produces:
      - "application/json"
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/CustomerResponse'
        404:
          description: "User not found"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"

securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  CustomerResponse:
    type: "object"
    properties:
      result:
        type: object
        properties:
          paging: 
            type: object
          results:
            type: array
            items:
              type: object
  CardTokenResponse:
    type: object
    properties:
      status:
        type: integer
      response:
        type: object
        properties:
          card_id:
            type: string
          id:
            type: string
  PaymentResponse:
    type: object
    properties:
      status:
        type: integer
      response:
        type: object
      

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"